+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 ## 1.0.8
 * Downgraded Edge DevTools version to 81.0.416 (Microsoft Edge Stable)
 * Merged Network Tool into the extension - [PR #128](https://github.com/microsoft/vscode-edge-devtools/pull/128)
-* Renaming tool to "Microsoft Edge (Chromium) Tools for VSCode"
+* Renaming tool to "Microsoft Edge Tools for VS Code"
 
 ## 1.0.7
 * Updated to version 82.0.423